Code clean up server.js Code clean up models.js
